backupnew/requirements.txt
# Core Framework
streamlit>=1.28.0
streamlit-aggrid>=0.3.4

# Computer Vision and ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
scikit-image
segment-anything>=1.0

# Vision Models
open-clip-torch>=2.20.0
transformers[torch]==4.38.2

# API and Web
requests>=2.31.0
aiohttp>=3.8.0
google-search-results>=2.4.2
openai>=1.0.0
replicate>=0.15.0
serpapi
httpx>=0.27.0

# Data Processing
pandas>=2.0.0
bottleneck>=1.5.0

# LangChain Components
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
pycocotools>=2.0.6

# Image Processing
imageio>=2.31.0
matplotlib>=3.7.0

# Caching and Performance
functools32; python_version < '3.2'
lru-dict>=1.2.0

# Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Additional Utilities
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.0.0
click>=8.0.0
numba<0.60.0
scipy<1.13.0
streamlit-drawable-canvas>=0.8.0

# CLIP from GitHub
git+https://github.com/openai/CLIP.git